Revert update to HEL file from theorists
I am really confused about what happened. I thought the
updated version was working (and I used it for the
`ttV/8/*` samples, but now it is complaining that "masses and
widths must be real" when I try to run a negative Wilson
coefficient. I'm reverting it for now until I have time to
investigate.