Parallelize range finding
A binomial search was being used previously to find the ranges, but that
only makes sense if CPU time is the bottleneck: this is not the case for
this workflow at the CRC. A parallelized binomial search would have
been more efficient but I decided to prioritize simplicity with this
basic approach which finds the range from a coarse scan, then runs a
finer scan.

Additionally, this revives the functionality to perform multidimensional
Wilson coefficient scans.